Refactor ConnectionInfoParameter to enums package
Move ConnectionInfoParameter class to a more appropriate package, rentasad.library.db.enums. This update improves code organization and complies with package naming conventions.
